Correcting the ranges for the ablation of attention and extra feature.
I think the ranges of deletion for ablation is wrong in Patrick s implementation.

The e2e paper mention that they use a concat of [start token, end token, head attention, extra feature] to represent span.

But the reality is that the span is constructed as [start token, end token,  extra feature, head attention,] in the code.